+++ magda-web-client/src/Components/Dataset/Add/Pages/DatasetAddEndPreviewPage.tsx
@@ -2,6 +2,7 @@ import React from "react";
 
 import { Link } from "react-router-dom";
 
+// If you are in Preview Mode
 export default function DatasetAddEndPreviewPage() {
     return (
         <div className="row people-and-production-page">