+++ CHANGES.md
@@ -43,6 +43,7 @@ UI:
 -   Added UI for "Publish as Open Data to data.gov.au"
 -   Refactor view dataset page & remove the edit function from the page
 -   Clarify tooltip text
+-   Use a new "All done!" screen at the end of the dataset flow
 -   Added Error Screen for adding distributions from URL feature & allow manually create distributions
 
 Storage: